modified all the sound related classes

+++ Template/Client/Include/Manager/Resource/SoundManager.h
@@ -52,7 +52,7 @@ public:
 	}
 
 	template <typename T>
-	void SetVolume(int volume)
+	void SetVolume(float volume)
 	{
 		SoundMap<T>& soundMap = GetSoundMap<T>();
